Some e2e test cleanup
