+++ test/e2e/utils/translationInfobar.ts
